backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Gmail API
google-api-python-client>=2.114.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# HTTP client
httpx>=0.26.0

# OpenAI
openai>=1.10.0

# Serialization
pydantic>=2.5.3
pydantic-settings>=2.1.0
orjson>=3.9.12

# Background tasks
apscheduler>=3.10.4

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
rich>=13.7.0
cryptography>=42.0.0
